pyproject.toml
[tool.poetry]
name = "aibackend"
version = "0.1.0"
description = ""
authors = [
    "AndresGM"
]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastai = ">=2.8.5,<3.0.0"
uvicorn = ">=0.38.0,<0.39.0"
redis = ">=7.0.1,<8.0.0"
langchain = ">=1.0.3,<2.0.0"
langchain-openai = ">=0.1.0,<1.0.0"
openai = ">=2.7.1,<3.0.0"
pydantic = ">=2.12.4,<3.0.0"
python-dotenv = ">=1.2.1,<2.0.0"
pytest = ">=8.4.2,<9.0.0"
pytest-asyncio = ">=1.2.0,<2.0.0"
httpx = ">=0.24.0,<1.0.0"

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"
